% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_function.R
\name{load_file_to_df}
\alias{load_file_to_df}
\title{Load a file into a data frame}
\usage{
load_file_to_df(file_path)
}
\arguments{
\item{file_path}{A string representing the file path to load.}
}
\value{
A data frame containing the contents of the file.
}
\description{
This function reads a CSV, Excel, or TXT file and loads it into a data frame.
The function automatically detects the file type based on the file extension.
}
\examples{
# Load a CSV file
df <- load_file_to_df("~/Desktop/ubc/stat 545/data/sample.csv")

# Load an Excel file
df <- load_file_to_df("~/Desktop/ubc/stat 545/data/sample.xlsx")

# Load a TXT file
df <- load_file_to_df("~/Desktop/ubc/stat 545/data/sample.txt")
}
